updated delete friend query

+++ YouGuessDataManipulationQueries.sql
@@ -5,7 +5,7 @@
 
 -- modifies friends table
 INSERT INTO Friends_List ('friendID') VALUES (:friendIDInput) WHERE `userID`= :userIDReference
-DELETE FROM Friends_List WHERE `friendID`= :friendIDReference
+DELETE FROM Friends_List WHERE `friendID`= :friendIDReference AND `userID`= :userIDReference
 
 -- modifies items table
 INSERT INTO Items_List (skipQuantity, reviveQuantity, userID) VALUES (:skipQuantity, :reviveQuantity, :userIDReference)
